Extracted components, getting past matches.
